refactor(card section): ♻️ have list and options as seperate components
